requirements.txt
setuptools~=65.5.1
pip~=22.3.1
wheel~=0.38.4
selenium
python-dateutil
pytz
